Revert "fix: support large models and partial biases"
This reverts commit dcb71977e83f0fcafbeb956bca51d2b6bc4a09ae.

+++ src/baremetal/model.h
@@ -1,13 +1,12 @@
 #ifndef BMT_MODEL_H
 #define BMT_MODEL_H
 
-#include <stddef.h>
 #include "baremetal.h"
 
 struct bm_model_s {
     bm_arch_t arch;
 
-    size_t   n_parameters;
+    int      n_parameters;
 
     int      kv_dim;
     int      head_size;
